# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetaRadar",
    "version": "5.1.0"
  },
  "paths": {
...
    "/api/v1/signals": {
      "get": {
        "tags": [
          "Signals & Intelligence"
        ],
        "summary": "List Signals",
        "description": "Returns filtered signals list with deterministic ordering, limit/offset pagination, and total count.",
        "operationId": "list_signals_api_v1_signals_get",
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by priority: CRITICAL, HIGH, MEDIUM, LOW",
              "title": "Severity"
            },
            "description": "Filter by priority: CRITICAL, HIGH, MEDIUM, LOW"
          },
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term in signal title or content",
              "title": "Entity"
            },
            "description": "Search term in signal title or content"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter signals published on or after date",
              "title": "Date From"
            },
            "description": "Filter signals published on or after date"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter signals published on or before date",
              "title": "Date To"
            },
            "description": "Filter signals published on or before date"
          },
          {
            "name": "signal_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by signal type",
              "title": "Signal Type"
            },
            "description": "Filter by signal type"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source ID",
              "title": "Source"
            },
            "description": "Filter by source ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/athena": {
      "post": {
        "tags": [
          "Signals & Intelligence"
        ],
        "summary": "Query Athena",
        "description": "Queries Athena intelligence synthesis layer with real pgvector retrieval, prompt sanitization, and honest citations.",
        "operationId": "query_athena_api_v1_athena_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AthenaQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AthenaQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AthenaQueryRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AthenaQueryRequest"
      },
      "AthenaQueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "confidence_type": {
            "type": "string",
            "title": "Confidence Type",
            "default": "model_reasoning"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "reasoning"
          },
          "model_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMetadataSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/AthenaEvidenceCitation"
            },
            "type": "array",
            "title": "Evidence"
          },
          "response_type": {
            "type": "string",
            "title": "Response Type",
            "default": "grounded_synthesis"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "confidence",
          "evidence_count"
        ],
        "title": "AthenaQueryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelMetadataSchema": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used",
            "default": false
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "reasoning_available": {
            "type": "boolean",
            "title": "Reasoning Available",
            "default": true
          },
          "actions_available": {
            "type": "boolean",
            "title": "Actions Available",
            "default": true
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "mode",
          "model"
        ],
        "title": "ModelMetadataSchema"
      },
...
      "SignalListResponse": {
        "properties": {
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalSchema"
            },
            "type": "array",
            "title": "Signals"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "signals",
          "total"
        ],
        "title": "SignalListResponse"
      },
      "SignalSchema": {
        "properties": {
          "signal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signal Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "development_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Id"
          },
          "pipeline_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Run Id"
          },
          "pmid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmid"
          },
          "nct_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nct Id"
          },
          "regulatory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulatory Id"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "signal_type": {
            "type": "string",
            "title": "Signal Type"
          },
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "ingested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingested At"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode",
            "default": "live"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": false
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "confidence_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Type"
          },
          "confidence_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Rationale"
          },
          "provenance_status": {
            "type": "string",
            "title": "Provenance Status",
            "default": "available"
          },
          "evidence_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Text"
          },
          "raw_record_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Record Reference"
          },
          "scoring_status": {
            "type": "string",
            "title": "Scoring Status",
            "default": "computed"
          },
          "facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Facts"
          },
          "interpretation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretation"
          },
          "speculation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speculation"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "MEDIUM"
          },
          "score_breakdown": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreBreakdownSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMetadataSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoring_model_version": {
            "type": "string",
            "title": "Scoring Model Version",
            "default": "haemophilia_v2.0"
          },
          "scoring_config_version": {
            "type": "string",
            "title": "Scoring Config Version",
            "default": "haemophilia_v1"
          },
          "embedding_model_version": {
            "type": "string",
            "title": "Embedding Model Version",
            "default": "v1"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "v1.0.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "signal_id",
          "source_id",
          "fingerprint",
          "signal_type",
          "disease",
          "title",
          "content",
          "published_at",
          "retrieved_at",
          "created_at"
        ],
        "title": "SignalSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}